refactor(logger): consolidate colors
